move type RpcImplConfiguration to cfx_rpc_cfx_types
