+++ crates/rpc/rpc-cfx-types/src/lib.rs
@@ -7,6 +7,7 @@ pub mod trace;
 pub mod trace_filter;
 pub mod traits;
 mod transaction_status;
+mod rpc_impl_configuration;
 
 pub use address::RpcAddress;
 pub use epoch_number::EpochNumber;
@@ -14,3 +15,4 @@ pub use fee_history::CfxFeeHistory;
 pub use fee_history_cache_entry::FeeHistoryCacheEntry;
 pub use phantom_block::PhantomBlock;
 pub use transaction_status::{PendingReason, TransactionStatus};
+pub use rpc_impl_configuration::RpcImplConfiguration;
